Move evoting testing utilities
