+++ fe1-web/src/features/evoting/__tests__/utils.ts
@@ -2,7 +2,7 @@ import { EventTags, Hash, Timestamp } from 'core/objects';
 import { CastVote, ElectionResult, EndElection } from 'features/evoting/network/messages';
 import STRINGS from 'resources/strings';
 import { mockLaoId, mockLaoIdHash } from '__tests__/utils';
-import { Election, ElectionStatus, Question, RegisteredVote, Vote } from '../Election';
+import { Election, ElectionStatus, Question, RegisteredVote, Vote } from '../objects/Election';
 
 const TIMESTAMP = new Timestamp(1609455600); // 1st january 2021
 const VERSION = STRINGS.election_version_identifier;
