+++ fe1-web/src/features/evoting/network/__tests__/ElectionMessageApi.test.ts
@@ -1,12 +1,10 @@
 import { publish } from 'core/network';
 import { channelFromIds, Timestamp } from 'core/objects';
+import { mockElectionNotStarted, mockElectionOpened } from 'features/evoting/__tests__/utils';
 import { SelectedBallots } from 'features/evoting/objects';
-import {
-  mockElectionNotStarted,
-  mockElectionOpened,
-} from 'features/evoting/objects/__tests__/utils';
 import 'jest-extended';
 import { mockLaoIdHash } from '__tests__/utils';
+
 import {
   castVote,
   openElection,
